refactor(aztec-node): extract NodePublicCallsSimulator class
Move the simulatePublicCalls implementation out of AztecNodeService into a
dedicated NodePublicCallsSimulator class. AztecNodeService now delegates to
it from a one-line method, preserving the @trackSpan decorator and the
public RPC interface. Move the eight associated unit tests into a new
test file that instantiates the simulator directly with mock dependencies.
