+++ yarn-project/aztec-node/src/aztec-node/server.ts
@@ -1,4 +1,4 @@
-import { Archiver, L1ToL2MessagesNotReadyError, createArchiver } from '@aztec/archiver';
+import { Archiver, createArchiver } from '@aztec/archiver';
 import { BBCircuitVerifier, BatchChonkVerifier, QueuedIVCVerifier } from '@aztec/bb-prover';
 import { TestCircuitVerifier } from '@aztec/bb-prover/test';
 import { type BlobClientInterface, createBlobClientWithFileStores } from '@aztec/blob-client/client';
@@ -20,7 +20,6 @@ import { retryUntil } from '@aztec/foundation/retry';
 import { count } from '@aztec/foundation/string';
 import { DateProvider, Timer } from '@aztec/foundation/timer';
 import { MembershipWitness, SiblingPath } from '@aztec/foundation/trees';
-import { isErrorClass } from '@aztec/foundation/types';
 import { type KeyStore, KeystoreManager, loadKeystores, mergeKeystores } from '@aztec/node-keystore';
 import { trySnapshotSync, uploadSnapshot } from '@aztec/node-lib/actions';
 import { createForwarderL1TxUtilsFromSigners, createL1TxUtilsFromSigners } from '@aztec/node-lib/factories';
@@ -40,8 +39,6 @@ import {
   SequencerClient,
   type SequencerPublisher,
 } from '@aztec/sequencer-client';
-import { buildCheckpointSimulationOverridesPlan } from '@aztec/sequencer-client/chain-state-overrides';
-import { PublicContractsDB, PublicProcessorFactory } from '@aztec/simulator/server';
 import {
   AttestationsBlockWatcher,
   BroadcastedInvalidCheckpointProposalWatcher,
@@ -50,7 +47,6 @@ import {
   type Watcher,
   createSlasher,
 } from '@aztec/slasher';
-import { CollectionLimitsConfig, PublicSimulatorConfig } from '@aztec/stdlib/avm';
 import { AztecAddress } from '@aztec/stdlib/aztec-address';
 import {
   type BlockData,
@@ -68,7 +64,6 @@ import {
   type CheckpointData,
   CheckpointReexecutionTracker,
   L1PublishedData,
-  type ProposedCheckpointData,
   type PublishedCheckpoint,
 } from '@aztec/stdlib/checkpoint';
 import type {
@@ -78,7 +73,6 @@ import type {
   NodeInfo,
   ProtocolContractAddresses,
 } from '@aztec/stdlib/contract';
-import { getNextL1SlotTimestamp, getSlotAtTimestamp, getTimestampForSlot } from '@aztec/stdlib/epoch-helpers';
 import { GasFees, type ManaUsageEstimate } from '@aztec/stdlib/gas';
 import { computePublicDataTreeLeafSlot } from '@aztec/stdlib/hash';
 import type {
@@ -109,14 +103,13 @@ import {
 } from '@aztec/stdlib/interfaces/server';
 import type { DebugLogStore, LogFilter, SiloedTag, Tag, TxScopedL2Log } from '@aztec/stdlib/logs';
 import { InMemoryDebugLogStore, NullDebugLogStore } from '@aztec/stdlib/logs';
-import { InboxLeaf, type L1ToL2MessageSource, appendL1ToL2MessagesToTree } from '@aztec/stdlib/messaging';
+import { InboxLeaf, type L1ToL2MessageSource } from '@aztec/stdlib/messaging';
 import type { Offense } from '@aztec/stdlib/slashing';
 import type { NullifierLeafPreimage, PublicDataTreeLeafPreimage } from '@aztec/stdlib/trees';
 import { MerkleTreeId, NullifierMembershipWitness, PublicDataWitness } from '@aztec/stdlib/trees';
 import {
   type FeeProvider,
   type GlobalVariableBuilder as GlobalVariableBuilderInterface,
-  GlobalVariables,
   type IndexedTxEffect,
   PublicSimulationOutput,
   type SimulationOverrides,
@@ -161,7 +154,7 @@ import {
 } from './block_response_helpers.js';
 import { type AztecNodeConfig, createKeyStoreForValidator } from './config.js';
 import { NodeMetrics } from './node_metrics.js';
-import { applyPublicDataOverrides } from './public_data_overrides.js';
+import { NodePublicCallsSimulator } from './node_public_calls_simulator.js';
 
 /**
  * The aztec node.
@@ -170,6 +163,7 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
   private metrics: NodeMetrics;
   // Prevent two snapshot operations to happen simultaneously
   private isUploadingSnapshot = false;
+  private readonly publicCallsSimulator: NodePublicCallsSimulator;
 
   public readonly tracer: Tracer;
 
@@ -205,6 +199,23 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
   ) {
     this.metrics = new NodeMetrics(telemetry, 'AztecNodeService');
     this.tracer = telemetry.getTracer('AztecNodeService');
+
+    this.publicCallsSimulator = new NodePublicCallsSimulator({
+      blockSource,
+      worldStateSynchronizer,
+      l1ToL2MessageSource,
+      contractDataSource,
+      globalVariableBuilder,
+      dateProvider,
+      telemetry,
+      l1ChainId,
+      config: {
+        rpcSimulatePublicMaxGasLimit: config.rpcSimulatePublicMaxGasLimit,
+        rpcSimulatePublicMaxDebugLogMemoryReads: config.rpcSimulatePublicMaxDebugLogMemoryReads,
+        rollupAddress: config.rollupAddress,
+      },
+      log: this.log,
+    });
 
     this.log.info(`Aztec Node version: ${this.packageVersion}`);
     this.log.info(`Aztec Node started on chain 0x${l1ChainId.toString(16)}`, pickL1ContractAddresses(config));
@@ -1470,222 +1481,12 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
   @trackSpan('AztecNodeService.simulatePublicCalls', (tx: Tx) => ({
     [Attributes.TX_HASH]: tx.getTxHash().toString(),
   }))
-  public async simulatePublicCalls(
+  public simulatePublicCalls(
     tx: Tx,
     skipFeeEnforcement = false,
     overrides?: SimulationOverrides,
   ): Promise<PublicSimulationOutput> {
-    // Check total gas limit for simulation
-    const gasSettings = tx.data.constants.txContext.gasSettings;
-    const txGasLimit = gasSettings.gasLimits.l2Gas;
-    const teardownGasLimit = gasSettings.teardownGasLimits.l2Gas;
-    if (txGasLimit + teardownGasLimit > this.config.rpcSimulatePublicMaxGasLimit) {
-      throw new BadRequestError(
-        `Transaction total gas limit ${
-          txGasLimit + teardownGasLimit
-        } (${txGasLimit} + ${teardownGasLimit}) exceeds maximum gas limit ${
-          this.config.rpcSimulatePublicMaxGasLimit
-        } for simulation`,
-      );
-    }
-
-    const txHash = tx.getTxHash();
-    const l2Tips = await this.blockSource.getL2Tips();
-    const latestBlockNumber = l2Tips.proposed.number;
-    const blockNumber = BlockNumber.add(latestBlockNumber, 1);
-
-    const hasProposedCheckpoint = l2Tips.proposedCheckpoint.checkpoint.number > l2Tips.checkpointed.checkpoint.number;
-    // True in two states: idle (no in-progress proposed checkpoint), or the latest proposed block is
-    // also the terminating block of a pending proposed checkpoint. False mid-checkpoint when blocks
-    // have been added past the last L1-confirmed boundary but no proposed-checkpoint entry exists yet.
-    const atCheckpointBoundary = l2Tips.proposedCheckpoint.block.number === l2Tips.proposed.number;
-
-    let newGlobalVariables: GlobalVariables;
-    let nextCheckpointMessages: Fr[] | undefined;
-    let targetCheckpoint: CheckpointNumber | undefined;
-
-    if (!atCheckpointBoundary) {
-      // Case A: continuation in an in-progress checkpoint. Every block in a checkpoint shares the
-      // same CheckpointGlobalVariables (slot, timestamp, gasFees, coinbase, feeRecipient), so we
-      // reuse the latest proposed block's globals verbatim and only bump the block number.
-      const latestBlockData = await this.blockSource.getBlockData({ number: latestBlockNumber });
-      if (!latestBlockData) {
-        throw new Error(`Latest proposed block ${latestBlockNumber} has no header on this node`);
-      }
-      newGlobalVariables = GlobalVariables.from({
-        ...latestBlockData.header.globalVariables,
-        blockNumber,
-      });
-      nextCheckpointMessages = undefined;
-    } else {
-      // Case B: opening a new checkpoint. Compute fresh globals.
-      // Slot is wall-clock anchored to match `FeeProviderImpl.computeCurrentMinFees` exactly so
-      // the simulator's `gasFees` agree with what the wallet observes via `getCurrentMinFees`.
-      // Mirroring the wallet's path is what keeps fee enforcement consistent across the two
-      // call sites; an archiver-anchored slot diverges whenever the local archiver lags wall-clock
-      // by an L2 slot, which is routine in e2e tests with short anvil intervals.
-      const rollupContract = this.globalVariableBuilder.getRollupContract();
-      const l1Constants = await this.blockSource.getL1Constants();
-      const lastCheckpointOnL1 = await rollupContract.getPendingCheckpoint();
-      const earliestTimestamp = getTimestampForSlot(SlotNumber.add(lastCheckpointOnL1.slotNumber, 1), l1Constants);
-      const nextEthTimestamp = getNextL1SlotTimestamp(this.dateProvider.nowInSeconds(), l1Constants);
-      const targetTimestamp = earliestTimestamp > nextEthTimestamp ? earliestTimestamp : nextEthTimestamp;
-      const targetSlot = getSlotAtTimestamp(targetTimestamp, l1Constants);
-
-      let proposedCheckpointData: ProposedCheckpointData | undefined;
-      let checkpointNumber: CheckpointNumber;
-      // We only build the overrides plan when extending a 1-deep proposed parent (i.e. the
-      // proposed parent's checkpoint number is exactly `checkpointed + 1`). Beyond that depth,
-      // the parent's grandparent has not landed on L1, so the helper would revert when reading
-      // the grandparent feeHeader. Without a proposed parent (or with a deeper pipeline), we
-      // omit the plan entirely so `getManaMinFeeAt` reads L1 state as-is — matching the fee
-      // value wallets see via `getCurrentMinFees`.
-      let buildOverridesForProposedParent = false;
-
-      if (hasProposedCheckpoint) {
-        // Extend the proposed parent: target checkpoint is parent + 1, and the parent's data feeds
-        // into the overrides plan so getManaMinFeeAt sees the chain state propose() will write.
-        proposedCheckpointData = await this.blockSource.getProposedCheckpointData();
-        const expectedNumber = l2Tips.proposedCheckpoint.checkpoint.number;
-        const expectedLastBlock = l2Tips.proposedCheckpoint.block.number;
-        const actualLastBlock = proposedCheckpointData
-          ? BlockNumber(proposedCheckpointData.startBlock + proposedCheckpointData.blockCount - 1)
-          : undefined;
-        if (
-          !proposedCheckpointData ||
-          proposedCheckpointData.checkpointNumber !== expectedNumber ||
-          actualLastBlock !== expectedLastBlock
-        ) {
-          throw new Error(
-            `Torn L2 tips snapshot: getL2Tips() reported proposed checkpoint ${expectedNumber} ending at block ${expectedLastBlock}, but getProposedCheckpointData() reported ${
-              proposedCheckpointData?.checkpointNumber ?? 'undefined'
-            } / ${actualLastBlock ?? 'undefined'}. Retry the simulation.`,
-          );
-        }
-        checkpointNumber = CheckpointNumber(proposedCheckpointData.checkpointNumber + 1);
-        // Only build the overrides plan when the proposed parent is 1-deep. When the pipeline is
-        // deeper (parent's checkpoint number > checkpointed + 1), the grandparent is not yet on
-        // L1 and `computePipelinedParentFeeHeader` would revert with
-        // `Rollup__UnavailableTempCheckpointLog`.
-        buildOverridesForProposedParent =
-          proposedCheckpointData.checkpointNumber === l2Tips.checkpointed.checkpoint.number + 1;
-      } else {
-        proposedCheckpointData = undefined;
-        checkpointNumber = CheckpointNumber(l2Tips.checkpointed.checkpoint.number + 1);
-      }
-
-      // Build the overrides plan only when extending a 1-deep proposed parent. Without it (idle
-      // case, or 2+ deep pipeline), we read L1 fees as-is and avoid the deep-pipeline revert,
-      // which also keeps `getManaMinFeeAt` in lockstep with the wallet's `getCurrentMinFees`.
-      const overridesPlan = buildOverridesForProposedParent
-        ? await buildCheckpointSimulationOverridesPlan({
-            checkpointNumber,
-            proposedCheckpointData,
-            invalidateToPendingCheckpointNumber: undefined,
-            checkpointedCheckpointNumber: l2Tips.checkpointed.checkpoint.number,
-            rollup: rollupContract,
-            signatureContext: { chainId: this.l1ChainId, rollupAddress: this.config.rollupAddress },
-            log: this.log,
-          })
-        : undefined;
-
-      // Simulation always zeroes these regardless of whether a sequencer is configured on this
-      // node — the simulator does not represent the proposer's payout addresses.
-      const coinbase = EthAddress.ZERO;
-      const feeRecipient = AztecAddress.ZERO;
-      const checkpointGlobals = await this.globalVariableBuilder.buildCheckpointGlobalVariables(
-        coinbase,
-        feeRecipient,
-        targetSlot,
-        overridesPlan,
-      );
-      newGlobalVariables = GlobalVariables.from({ blockNumber, ...checkpointGlobals });
-      targetCheckpoint = checkpointNumber;
-    }
-
-    const publicProcessorFactory = new PublicProcessorFactory(
-      this.contractDataSource,
-      new DateProvider(),
-      this.telemetry,
-      this.log.getBindings(),
-    );
-
-    this.log.verbose(`Simulating public calls for tx ${txHash}`, {
-      globalVariables: newGlobalVariables.toInspect(),
-      txHash,
-      blockNumber,
-      atCheckpointBoundary,
-      hasProposedCheckpoint,
-    });
-
-    // Ensure world-state has caught up with the latest block we loaded from the archiver before
-    // fetching L1-to-L2 messages or forking, so all reads observe a coherent view.
-    await this.worldStateSynchronizer.syncImmediate(latestBlockNumber);
-
-    // Fetch L1-to-L2 messages only when opening a new checkpoint, after sync so the archiver
-    // has had a chance to advance to the tip we observed.
-    if (targetCheckpoint !== undefined) {
-      nextCheckpointMessages = await this.l1ToL2MessageSource.getL1ToL2Messages(targetCheckpoint).catch(err => {
-        if (isErrorClass(err, L1ToL2MessagesNotReadyError)) {
-          this.log.warn(
-            `L1-to-L2 messages for checkpoint ${targetCheckpoint} are not ready yet (simulating without them)`,
-          );
-        } else {
-          this.log.error(
-            `Failed to get L1-to-L2 messages for checkpoint ${targetCheckpoint} (simulating without them)`,
-            err,
-          );
-        }
-        return undefined;
-      });
-    }
-
-    // Request a new fork of the world state at the latest block number, and apply any overrides and next checkpoint messages to it before simulation
-    await using merkleTreeFork = await this.worldStateSynchronizer.fork(latestBlockNumber);
-
-    if (nextCheckpointMessages !== undefined) {
-      this.log.debug(
-        `Appending ${nextCheckpointMessages.length} L1-to-L2 messages to the world state tree for the next checkpoint`,
-        { checkpointNumber: targetCheckpoint },
-      );
-      await appendL1ToL2MessagesToTree(merkleTreeFork, nextCheckpointMessages);
-    }
-    await applyPublicDataOverrides(merkleTreeFork, overrides?.publicStorage);
-
-    const config = PublicSimulatorConfig.from({
-      skipFeeEnforcement,
-      collectDebugLogs: true,
-      collectHints: false,
-      collectCallMetadata: true,
-      collectStatistics: false,
-      collectionLimits: CollectionLimitsConfig.from({
-        maxDebugLogMemoryReads: this.config.rpcSimulatePublicMaxDebugLogMemoryReads,
-      }),
-    });
-
-    const contractsDB = new PublicContractsDB(this.contractDataSource, this.log.getBindings());
-    if (overrides?.contracts) {
-      contractsDB.addContracts(Object.values(overrides.contracts).map(({ instance }) => instance));
-    }
-    const processor = publicProcessorFactory.create(merkleTreeFork, newGlobalVariables, config, contractsDB);
-
-    // REFACTOR: Consider merging ProcessReturnValues into ProcessedTx
-    const [processedTxs, failedTxs, _usedTxs, returns, debugLogs] = await processor.process([tx]);
-    // REFACTOR: Consider returning the error rather than throwing
-    if (failedTxs.length) {
-      this.log.warn(`Simulated tx ${txHash} fails: ${failedTxs[0].error}`, { txHash });
-      throw failedTxs[0].error;
-    }
-
-    const [processedTx] = processedTxs;
-    return new PublicSimulationOutput(
-      processedTx.revertReason,
-      processedTx.globalVariables,
-      processedTx.txEffect,
-      returns,
-      processedTx.gasUsed,
-      debugLogs,
-    );
+    return this.publicCallsSimulator.simulate(tx, skipFeeEnforcement, overrides);
   }
 
   public async isValidTx(
